break image_io into image_io and image
